Fix arc42/GDD compliance violations (excluding Consequence/Requirement)
HIGHLANDER Template/Instance Separation:
- Split EmergencySituation into immutable template + ActiveEmergencyState
- Split ObservationScene into immutable template + ObservationSceneState
- Move ObligationActivity pending results to GameWorld (service now stateless)
- Update EmergencyFacade, ObservationFacade, GameFacade, GameScreen

Backend/Frontend Separation:
- Remove FocusClass, WeatherIcon presentation properties from ViewModels
- Remove Icon from RouteTokenRequirementViewModel, add TokenType enum
- Frontend now computes presentation from domain values

Type System Compliance:
- MarketPriceInfo.SupplyLevel: float → int (0-100 percentage)
- CardAnimationState.AnimationDelay: double → int AnimationDelayMs
- PathfindingResult.TotalCost: internal float → public int (API boundary)

Entity Model Compliance:
- Replace FamiliarityEntry with typed RouteFamiliarityEntry/LocationFamiliarityEntry
- Use object references (Route, Location) instead of string EntityId
- Update Player methods to use object equality not string comparison

Files changed: 15
All fixes per audit_reports/ findings, excluding Consequence/Requirement system.

+++ src/Models/Market/MarketPriceInfo.cs
@@ -1,6 +1,7 @@
 /// <summary>
 /// Market price information for an item at a location
 /// HIGHLANDER: Object references, no string IDs
+/// TYPE SYSTEM: int only for game values (no float/double)
 /// </summary>
 public class MarketPriceInfo
 {
@@ -9,7 +10,7 @@ public class MarketPriceInfo
     public Item Item { get; set; }
     public int BuyPrice { get; set; }
     public int SellPrice { get; set; }
-    public float SupplyLevel { get; set; }
+    public int SupplyLevel { get; set; } // 0-100 percentage
     public bool IsCurrentLocation { get; set; }
     public bool CanBuy { get; set; }
 }
